restructure crag and route page
