+++ src/app/[lang]/crag/[cragSlug]/(crag)/comments/components/add-comment-form.tsx
@@ -1,12 +1,15 @@
 "use client";
 
 import { useState } from "react";
-import Button from "../../../../../../components/ui/button";
-import TextArea from "../../../../../../components/ui/text-area";
-import { Radio, RadioGroup } from "../../../../../../components/ui/radio-group";
+import Button from "../../../../../../../components/ui/button";
+import TextArea from "../../../../../../../components/ui/text-area";
+import {
+  Radio,
+  RadioGroup,
+} from "../../../../../../../components/ui/radio-group";
 import createCommentAction from "./server-actions/create-comment-action";
 import { useRouter } from "next/navigation";
-import { User } from "../../../../../../graphql/generated";
+import { User } from "../../../../../../../graphql/generated";
 
 interface Props {
   cragId: string;
