+++ src/app/[lang]/crag/[cragSlug]/(crag)/comments/components/comment.tsx
@@ -1,4 +1,4 @@
-import { User } from "../../../../../../graphql/generated";
+import { User } from "../../../../../../../graphql/generated";
 import CommentActions from "./comment-actions";
 
 interface CommentProps {
