+++ src/app/[lang]/crag/[cragSlug]/(crag)/comments/components/server-actions/delete-comment-action.tsx
@@ -1,8 +1,8 @@
 "use server";
 
 import { gql } from "urql/core";
-import { DeleteCommentDocument } from "../../../../../../../graphql/generated";
-import urqlServer from "../../../../../../../graphql/urql-server";
+import { DeleteCommentDocument } from "../../../../../../../../graphql/generated";
+import urqlServer from "../../../../../../../../graphql/urql-server";
 
 async function deleteCommentAction(commentId: string) {
   const result = await urqlServer().mutation(DeleteCommentDocument, {
