+++ src/app/[lang]/crag/[cragSlug]/(crag)/comments/page.tsx
@@ -1,9 +1,12 @@
 import { gql } from "urql/core";
-import urqlServer from "../../../../../graphql/urql-server";
-import { Crag, CragCommentsDocument } from "../../../../../graphql/generated";
+import urqlServer from "../../../../../../graphql/urql-server";
+import {
+  Crag,
+  CragCommentsDocument,
+} from "../../../../../../graphql/generated";
 import Comment, { CommentType } from "./components/comment";
 import AddCommentForm from "./components/add-comment-form";
-import authStatus from "../../../../../utils/auth/auth-status";
+import authStatus from "../../../../../../utils/auth/auth-status";
 
 interface Params {
   cragSlug: string;
