+++ src/app/[lang]/crag/[cragSlug]/(crag)/components/crag-gallery.tsx
