+++ src/app/[lang]/crag/[cragSlug]/(crag)/components/crag-header.tsx
@@ -1,5 +1,8 @@
-import { Crag } from "../../../../../graphql/generated";
-import { Breadcrumb, Breadcrumbs } from "../../../../../components/breadcrumbs";
+import { Crag } from "../../../../../../graphql/generated";
+import {
+  Breadcrumb,
+  Breadcrumbs,
+} from "../../../../../../components/breadcrumbs";
 import CragTabMenu from "./crag-header/crag-tab-menu";
 
 interface Props {
