+++ src/app/[lang]/crag/[cragSlug]/(crag)/components/crag-info.tsx
