+++ src/app/[lang]/crag/[cragSlug]/(crag)/components/crag-routes/crag-route-list.tsx
@@ -1,10 +1,10 @@
 import { useContext } from "react";
-import { Crag, Route } from "../../../../../../graphql/generated";
+import { Crag, Route } from "../../../../../../../graphql/generated";
 import CragRoute, { CragRouteCompact } from "./crag-route-list/crag-route";
 import { CragRoutesContext, FilterOptions, SortOptions } from "../crag-routes";
-import IconStarFull from "../../../../../../components/ui/icons/star-full";
-import IconComment from "../../../../../../components/ui/icons/comment";
-import IconCheck from "../../../../../../components/ui/icons/check";
+import IconStarFull from "../../../../../../../components/ui/icons/star-full";
+import IconComment from "../../../../../../../components/ui/icons/comment";
+import IconCheck from "../../../../../../../components/ui/icons/check";
 import { IconSize } from "@/components/ui/icons/icon-size";
 import { filterEntitiesBySearchTerm } from "@/utils/search-helpers";
