+++ src/app/[lang]/crag/[cragSlug]/(crag)/components/crag-routes/crag-route-list/crag-route/route-grade.tsx
@@ -1,6 +1,6 @@
 import Dialog, {
   DialogSize,
-} from "../../../../../../../../components/ui/dialog";
+} from "../../../../../../../../../components/ui/dialog";
 import DifficultyVotes from "./difficulty-votes";
 import { Crag, DifficultyVote, Route } from "@/graphql/generated";
 import Grade from "@/components/grade";
