+++ src/app/[lang]/crag/[cragSlug]/(crag)/components/crag-routes/crag-route-list/crag-route/server-actions/difficulty-votes-action.ts
