+++ src/app/[lang]/crag/[cragSlug]/(crag)/components/crag-routes/crag-sector.tsx
@@ -1,5 +1,5 @@
-import { Crag, Sector } from "../../../../../../graphql/generated";
-import Accordion from "../../../../../../components/ui/accordion";
+import { Crag, Sector } from "../../../../../../../graphql/generated";
+import Accordion from "../../../../../../../components/ui/accordion";
 import CragRouteList from "./crag-route-list";
 
 interface Props {
