+++ src/app/[lang]/crag/[cragSlug]/(crag)/gallery/components/image-list/image-list-slide.tsx
