+++ src/app/[lang]/crag/[cragSlug]/(crag)/gallery/components/image-list/index.tsx
