+++ src/app/[lang]/crag/[cragSlug]/(crag)/gallery/page.tsx
