+++ src/app/[lang]/crag/[cragSlug]/(crag)/info/page.tsx
