+++ src/app/[lang]/crag/[cragSlug]/(crag)/page.tsx
@@ -4,10 +4,10 @@ import {
   Crag,
   CragSectorsDocument,
   MyCragSummaryDocument,
-} from "../../../../graphql/generated";
-import urqlServer from "../../../../graphql/urql-server";
+} from "../../../../../graphql/generated";
+import urqlServer from "../../../../../graphql/urql-server";
 import CragRoutes from "./components/crag-routes";
-import authStatus from "../../../../utils/auth/auth-status";
+import authStatus from "../../../../../utils/auth/auth-status";
 
 type Params = {
   cragSlug: string;
